+++ src/restore.js
@@ -1,6 +1,6 @@
 import * as cache from "@actions/cache";
 import * as core from "@actions/core";
-import { CACHE_KEY_PREFIX, getCacheRestorePath, getExistingCacheEntries } from "./helpers.js";
+import { CACHE_KEY_PREFIX, deleteCacheEntry, getCacheRestorePath, getExistingCacheEntries } from "./helpers.js";
 
 const token = core.getInput("token");
 const vcpkgArchivePath = core.getInput("archive-path");
@@ -18,7 +18,14 @@ await core.group("Restoring vcpkg cache", async () => {
       const archivePath = getCacheRestorePath(vcpkgArchivePath, cacheKey);
       core.info(`Restoring '${cacheKey}' to '${archivePath}'`);
 
-      await cache.restoreCache([archivePath], cacheKey);
+      const savedCacheKey = await cache.restoreCache([archivePath], cacheKey, undefined, undefined, true);
+      if (!savedCacheKey) {
+        core.warning(
+          "Cache failed to restore (likely due to a version mismatch from using a different archive path). Deleting old entry"
+        );
+
+        await deleteCacheEntry(token, cacheKey);
+      }
     })
   );
 });